[Refactor] refactor counselorPage components
