+++ src/modules/counselorPage/CounselorVideo.tsx
@@ -1,0 +1,9 @@
+export default function CounselorVideo() {
+  return (
+    <div className="py-12 lg:py-14">
+      <div className="container h-[212px] lg:h-[276px]">
+        <iframe className="h-full w-full" src="https://www.youtube.com/embed/qpOcRG3e9Q8" title="YouTube video player" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" />
+      </div>
+    </div>
+  );
+}
