+++ src/pages/counselorlist/counselorpage/[id].tsx
@@ -1,17 +1,22 @@
-/* eslint-disable react/no-unused-prop-types */
 import { useState } from 'react';
 import Image from 'next/image';
 import { useRouter } from 'next/router';
-import { Breadcrumb, Collapse, ConfigProvider, Radio, RadioChangeEvent, Select } from 'antd';
+import { ConfigProvider, Radio, RadioChangeEvent, Select } from 'antd';
 import axios from 'axios';
 import { getCookie } from 'cookies-next';
 import useCloseLoading from '@/common/hooks/useCloseLoading';
+import useOpenLoading from '@/common/hooks/useOpenLoading';
 import { useAddToCartPostMutation } from '@/common/redux/service/counselorPage';
 import { IButton } from '@/common/components/IButton';
 import { counselorPageBreadcrumb } from '@/lib/counselorPage/CounselorPageData';
 import checkCircle from 'public/images/check-circle.svg';
-import rateStar from 'public/images/rateStar.svg';
 import convertFieldId from '@/common/helpers/convertFieldId';
+import RegularQuestion from '@/modules/counselorPage/RegularQuestion';
+import UserComment from '@/modules/counselorPage/UserComment';
+import CounselorVideo from '@/modules/counselorPage/CounselorVideo';
+import CounselorRate from '@/modules/counselorPage/CounselorRate';
+import CounselorInformation from '@/modules/counselorPage/CounselorInformation';
+import { ICounselorPageProps, ICourses, IFilterCases } from '@/types/interface';
 
 // 使用axios取得path
 export const getServerSidePaths = async () => {
@@ -38,53 +43,6 @@ export const getServerSideProps = async ({ params }: { params: { id: string } })
   };
 };
 
-// 折疊元件
-const { Panel } = Collapse;
-// 常見問題的資料陣列
-const questionData = [
-  {
-    question: '預約方式',
-    answer: (
-      <ul className=" font-normal text-gray-900">
-        <li>1. 選擇心儀的課程方案後，點選「手刀預約」</li>
-        <li>2. 前往購物車完成結帳</li>
-        <li>3. 前往 會員中心 / 預約管理 / 待預約 選擇預約時段</li>
-        <li>4. 等待諮商師接受預約，收到接受通知後就完成囉！</li>
-      </ul>
-    ),
-  },
-  {
-    question: '費用說明',
-    answer: <p className=" font-normal text-gray-900">每堂課皆為一小時，可以自行選擇預約堂數，單堂費用由諮商師自行訂定，不同諮商主題的費用可能不同。</p>,
-  },
-  {
-    question: '上課說明',
-    answer: <p className=" font-normal text-gray-900">課程將透過 ZOOM 線上進行，預約時間十分鐘前會在會員中心釋出課程連結，只要點選連結，就可以開始上課囉！</p>,
-  },
-  {
-    question: '退課須知',
-    answer: <p className=" font-normal text-gray-900">預約成功後若要辦理退課，請聯絡客服信箱由小幫手協助處理。提醒：為維護雙方權益，請審慎考慮後再申請退課。</p>,
-  },
-];
-
-interface ICounselorPageProps {
-  Data: {
-    Name: string;
-    FieldTags: string[];
-    Photo: string;
-    SelfIntroduction: string;
-    Fields: any;
-  };
-}
-interface ICourses {
-  Item: string;
-  Price: number;
-}
-interface IFilterCases {
-  label: string;
-  value: number;
-}
-
 // Server 渲染步驟 ＝ 渲染諮商師個人資料 => 渲染諮商師的專長領域選項 => 渲染專長領域的說明文字 => 渲染諮商師的課程方案
 // Client 互動步驟 ＝ 選擇專長領域 => 選擇課程方案 => 加入購物車
 
@@ -92,13 +50,14 @@ export default function CounselorPage({ data, counselorId }: { data: ICounselorP
   // ==================== 關閉 loading ====================
   useCloseLoading();
 
+  const openLoading = useOpenLoading();
   const [addToCartPost] = useAddToCartPostMutation();
   const token = getCookie('auth');
   const router = useRouter();
   const { Name, FieldTags, Photo, SelfIntroduction, Fields } = data.Data;
   const [chooseCase, setChooseCase] = useState(null);
 
-  // ====================Server 渲染畫面====================
+  // ==================== Server 渲染畫面 ====================
   // 取得專長領域的選項
   const FieldOptions = Fields.map(({ Field }: { Field: string }) => ({ label: Field, value: Field }));
 
@@ -131,11 +90,25 @@ export default function CounselorPage({ data, counselorId }: { data: ICounselorP
   // 渲染課程方案的選項
   const [chooseCourse, setChooseCourse] = useState(topicOptions[0]);
 
-  // ====================Client 畫面互動====================
+  // ==================== Client 畫面互動 ====================
 
   // 課程方案篩選
   const filterCase = (value: string) => Fields.flatMap((item: { Courses: ICourses[]; Field: string }) => {
-    const convertData = item.Courses.filter(() => value === item.Field).map((item2: { Item: string; Price: number }) => ({ label: `${item2.Item} / ${item2.Price} 元`, value: item2.Item }));
+    const convertData = item.Courses.filter(() => value === item.Field).map(({ Item, Price }) => {
+      if (Item === '體驗課一堂') {
+        return { label: `體驗課 60分鐘 / ${Price.toLocaleString()} 元`, value: Item };
+      }
+      if (Item === '一堂') {
+        return { label: `${Item} 60分鐘 / ${Price.toLocaleString()} 元`, value: Item };
+      }
+      if (Item === '三堂') {
+        return { label: `${Item} 3小時 / ${Price.toLocaleString()} 元`, value: Item };
+      }
+      if (Item === '五堂') {
+        return { label: `${Item} 5小時 / ${Price.toLocaleString()} 元`, value: Item };
+      }
+      return { label: `${Item} / ${Price.toLocaleString()} 元`, value: Item };
+    });
     return convertData;
   });
 
@@ -185,7 +158,7 @@ export default function CounselorPage({ data, counselorId }: { data: ICounselorP
       alert('請選擇方案');
       return;
     }
-
+    openLoading();
     const FieldId = convertFieldId(chooseTopic);
     const CounselorId = Number(counselorId);
 
@@ -201,6 +174,7 @@ export default function CounselorPage({ data, counselorId }: { data: ICounselorP
       const {
         data: { Message },
       } = res.error as { data: { Message: string } };
+      useCloseLoading();
       alert(Message);
     }
 
@@ -210,6 +184,7 @@ export default function CounselorPage({ data, counselorId }: { data: ICounselorP
       router.push('/shoppingcart');
       alert(resData.Message);
     } else {
+      useCloseLoading();
       alert(resData?.Message || '加入購物車失敗');
     }
   };
@@ -217,53 +192,12 @@ export default function CounselorPage({ data, counselorId }: { data: ICounselorP
   return (
     <>
       {/* 諮商師資料 */}
-      <section className="bg-primary py-14 lg:pt-[84px] lg:pb-[124px]">
-        <div className="container">
-          <Breadcrumb items={counselorPageBreadcrumb} />
-          <div className="mt-6 flex w-full justify-center lg:mt-14">
-            <div className="flex flex-col items-center lg:w-full lg:max-w-[1012px] lg:flex-row lg:items-center lg:justify-between">
-              <Image className="rounded-2xl" src={Photo} alt={Name} width={400} height={400} priority />
-
-              <div className="mt-10 w-full max-w-[340px] border-y border-secondary pt-6 pb-8 lg:mt-0 lg:min-h-[400px] lg:max-w-[492px] lg:pt-10 lg:pb-[45px]">
-                <h2 className="mb-4 w-full text-left lg:mb-6">{Name}</h2>
-
-                {/* 手機版FieldTags */}
-                <ul className="mb-8 flex w-[340px] flex-wrap lg:hidden">
-                  {FieldTags.map((topic: string, index: number) => {
-                    if (index < 3) {
-                      return <li className="fakeBorder mr-4 w-full max-w-[96px] rounded-full py-3 text-center text-sm font-semibold text-secondary ">{topic}</li>;
-                    }
-                    if (index >= 3) {
-                      return <li className="fakeBorder mr-4  mt-3 w-full max-w-[96px] rounded-full py-3 text-center text-sm font-semibold text-secondary ">{topic}</li>;
-                    }
-                    return null;
-                  })}
-                </ul>
-
-                {/* 電腦版FieldTags */}
-                <ul className="hidden flex-wrap  lg:mb-14 lg:flex">
-                  {FieldTags.map((topic: string, index: number) => {
-                    if (index < 4) {
-                      return <li className="fakeBorder mr-3 w-full rounded-full py-3 text-center text-sm  font-semibold text-secondary lg:max-w-[104px]">{topic}</li>;
-                    }
-                    if (index >= 4) {
-                      return <li className="fakeBorder mt-3  mr-3 w-full rounded-full py-3 text-center text-sm  font-semibold text-secondary lg:max-w-[104px]">{topic}</li>;
-                    }
-                    return null;
-                  })}
-                </ul>
-
-                {/* 自我介紹 */}
-                <p className="text-sm text-gray-900 lg:text-lg">{SelfIntroduction}</p>
-              </div>
-            </div>
-          </div>
-        </div>
-      </section>
+      <CounselorInformation counselorPageBreadcrumb={counselorPageBreadcrumb} Photo={Photo} Name={Name} SelfIntroduction={SelfIntroduction} FieldTags={FieldTags} />
 
       {/* 預約課程 */}
       <section className="lg:container lg:flex lg:justify-between lg:py-[148px]">
-        <div className="">
+        {/* 手機版 */}
+        <div>
           {/* 預約課程 */}
           <div className="py-20 lg:pt-0 lg:pb-14">
             <div className="container">
@@ -345,7 +279,7 @@ export default function CounselorPage({ data, counselorId }: { data: ICounselorP
                       return (
                         <li className="flex max-w-[340px] items-center space-x-3 lg:max-w-none">
                           <Image src={checkCircle} alt="checkCircle_icon" width={17.5} height={17.5} />
-                          <p className="text-gray-900">想要改善伴侶間爭吵、衝突的你們</p>
+                          <p className="text-gray-900">{featureTxt}</p>
                         </li>
                       );
                     }
@@ -412,46 +346,13 @@ export default function CounselorPage({ data, counselorId }: { data: ICounselorP
           </div>
 
           {/* 影片區塊 */}
-          <div className="py-12 lg:py-14">
-            <div className="container h-[212px] lg:h-[276px]">
-              <iframe className="h-full w-full" src="https://www.youtube.com/embed/qpOcRG3e9Q8" title="YouTube video player" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" />
-            </div>
-          </div>
+          <CounselorVideo />
 
           {/* 評分區塊 */}
-          <div className="container">
-            <div className="flex flex-col items-center border-t border-secondary py-12 lg:py-14">
-              <h2 className="mb-7 text-center lg:w-full lg:text-left lg:text-lg">諮商師評論數據</h2>
-
-              <ul className="flex w-full max-w-[308px] py-3 lg:max-w-[356px]">
-                <li className="w-1/3 font-bold text-secondary">
-                  <p className="mb-1 lg:text-2xl">
-                    99
-                    <span className="text-gray-500">％</span>
-                  </p>
-                  <p>滿意度</p>
-                </li>
-
-                <li className="w-1/3 min-w-[112px] border-x border-secondary text-center font-bold text-secondary">
-                  <p className="mb-1 lg:text-2xl">
-                    100
-                    <span className="text-gray-500">％</span>
-                  </p>
-                  <p>出席率</p>
-                </li>
-
-                <li className="w-1/3 text-end font-bold text-secondary">
-                  <p className="mb-1 lg:text-2xl">
-                    100
-                    <span className="text-gray-500">＋</span>
-                  </p>
-                  <p>個案人數</p>
-                </li>
-              </ul>
-            </div>
-          </div>
+          <CounselorRate />
         </div>
 
+        {/* 電腦版方案選擇 */}
         <div className="hidden lg:block lg:pt-[146px]">
           {/* 價格區塊 */}
           <div className="relative w-full rounded-2xl border-2 border-gray-700 bg-gray-100  pt-[60px] pb-12 lg:max-w-[388px] lg:pt-[78px] lg:pb-14">
@@ -505,152 +406,10 @@ export default function CounselorPage({ data, counselorId }: { data: ICounselorP
       </section>
 
       {/* 用戶好評 */}
-      <section className="bg-primary py-20 lg:py-[148px]">
-        <div className="container flex flex-col items-center">
-          <h2 className="mb-[46px] w-full text-center lg:mb-[72px] lg:text-left ">用戶好評</h2>
-
-          <ul className="lg:flex lg:w-full lg:justify-between">
-            <li>
-              <div className="flex h-[338px] w-[284px] flex-col justify-between rounded-[20px] bg-white py-12 px-6 text-gray-900 shadow-md">
-                <div>
-                  <ul className="mb-3 flex">
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                  </ul>
-
-                  <p>平台有心理師的簡介和評價，讓人更有方向去尋找。謝謝你們，讓遠在美國的我還可以找得到適合自己的心理師。</p>
-                </div>
-
-                <h3 className="text-xl font-bold">菲小姐</h3>
-              </div>
-            </li>
-
-            <li className="hidden lg:block">
-              <div className="flex h-[338px] w-[284px] flex-col justify-between rounded-[20px] bg-white py-12 px-6 text-gray-900 shadow-md">
-                <div className="">
-                  <ul className="mb-3 flex">
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                  </ul>
-
-                  <p className="mb-12">平台有心理師的簡介和評價，讓人更有方向去尋找。謝謝你們，讓遠在美國的我還可以找得到適合自己的心理師。</p>
-                </div>
-
-                <h3 className="mb-2 text-xl font-bold">菲小姐</h3>
-              </div>
-            </li>
-
-            <li className="hidden lg:block">
-              <div className="flex h-[338px] w-[284px] flex-col justify-between rounded-[20px] bg-white py-12 px-6 text-gray-900 shadow-md">
-                <div className="">
-                  <ul className="mb-3 flex">
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                  </ul>
-
-                  <p className="mb-12">平台有心理師的簡介和評價，讓人更有方向去尋找。謝謝你們，讓遠在美國的我還可以找得到適合自己的心理師。</p>
-                </div>
-
-                <h3 className="mb-2 text-xl font-bold">菲小姐</h3>
-              </div>
-            </li>
-
-            <li className="hidden lg:block">
-              <div className="flex h-[338px] w-[284px] flex-col justify-between rounded-[20px] bg-white py-12 px-6 text-gray-900 shadow-md">
-                <div className="">
-                  <ul className="mb-3 flex">
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                    <li>
-                      <Image src={rateStar} alt="rateStar" />
-                    </li>
-                  </ul>
-
-                  <p className="mb-12">平台有心理師的簡介和評價，讓人更有方向去尋找。謝謝你們，讓遠在美國的我還可以找得到適合自己的心理師。</p>
-                </div>
-
-                <h3 className="mb-2 text-xl font-bold">菲小姐</h3>
-              </div>
-            </li>
-          </ul>
-        </div>
-      </section>
+      <UserComment />
 
       {/* 常見問題 */}
-      <section className="bg-primary-tint py-20">
-        <div className="container lg:flex lg:max-w-[860px] lg:flex-col lg:items-center lg:px-0">
-          <h2 className="mb-7 w-full text-center lg:mb-14">常見問題</h2>
-
-          <div className="counselorPageQuestion w-full border-y border-gray-900">
-            <ConfigProvider
-              theme={{
-                token: {
-                  colorTextBase: '#424242',
-                  // 變更標題色
-                  colorBorder: '#424242',
-                },
-              }}
-            >
-              <Collapse bordered={false} expandIconPosition="end" className="bg-inherit">
-                {questionData.map(({ question, answer }) => (
-                  <Panel className="p-2 text-lg font-bold " header={question} key={question}>
-                    {answer}
-                  </Panel>
-                ))}
-              </Collapse>
-            </ConfigProvider>
-          </div>
-        </div>
-      </section>
+      <RegularQuestion />
     </>
   );
 }
